chore: 채팅 MessageErrorResponse 상태 추가
